pyproject.toml
[project]
name = "sec-blockchain-papers"
version = "0.1.0"
description = "Track blockchain-relevant papers from NDSS / USENIX Security / IEEE S&P / ACM CCS via DBLP."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "lxml>=5.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
